pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expense-manager"
version = "0.1.0"
description = "A simple interface to get insights from expenses"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "streamlit>=1.27.0",
    "sqlite-utils>=3.36",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    "plotly>=5.14.0",
    "pydantic[email]>=2.0.0",
    "email-validator>=2.0.0",
    "pandas-stubs>=2.2.3.250308",
]

[project.scripts]
run = "expense_manager.cli:run_app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.9.9",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "loguru>=0.7.3",
]

[tool.setuptools]
packages = ["expense_manager"]

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["E", "F", "I", "D"] # Error, Flake8, Import sorting, Docstrings
lint.ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
